Correction et suppression doublon
